// src/app/models/from-operating-room-to.model.ts
import { UCI } from './uci.model';
import { Recover } from './recover.model';

export class FromOperatingRoomTo {

  to: string;
  recover: Recover;
  uci: UCI;
  decease: Date;
  simpleDeceaseDate: string;
  simpleDeceaseHour: string;
  checkDate: Date;
  simpleCheckDate: string;
  simpleCheckHour: string;
  status: string;

}